refactor + style: accordion component
